+++ src/Access/ya.make
@@ -43,6 +43,7 @@ SRCS(
     SettingsProfile.cpp
     SettingsProfileElement.cpp
     SettingsProfilesCache.cpp
+    SettingsProfilesInfo.cpp
     User.cpp
     UsersConfigAccessStorage.cpp
     tests/gtest_access_rights_ops.cpp
