+++ src/Client/ya.make
@@ -12,9 +12,11 @@ PEERDIR(
 SRCS(
     Connection.cpp
     ConnectionEstablisher.cpp
+    ConnectionPool.cpp
     ConnectionPoolWithFailover.cpp
     HedgedConnections.cpp
     HedgedConnectionsFactory.cpp
+    IConnections.cpp
     MultiplexedConnections.cpp
 
 )
