+++ src/DataStreams/PushingToSinkBlockOutputStream.h
@@ -5,6 +5,11 @@
 namespace DB
 {
 
+namespace ErrorCodes
+{
+    extern const int LOGICAL_ERROR;
+}
+
 class PushingToSinkBlockOutputStream : public IBlockOutputStream
 {
 public:
