+++ src/Storages/MergeTree/MergeTreeThreadSelectProcessor.h
@@ -11,10 +11,10 @@ class MergeTreeReadPool;
 /** Used in conjunction with MergeTreeReadPool, asking it for more work to do and performing whatever reads it is asked
   * to perform.
   */
-class MergeTreeThreadSelectBlockInputProcessor : public MergeTreeBaseSelectProcessor
+class MergeTreeThreadSelectProcessor : public MergeTreeBaseSelectProcessor
 {
 public:
-    MergeTreeThreadSelectBlockInputProcessor(
+    MergeTreeThreadSelectProcessor(
         const size_t thread_,
         const std::shared_ptr<MergeTreeReadPool> & pool_,
         const size_t min_marks_to_read_,
@@ -32,7 +32,7 @@ public:
 
     String getName() const override { return "MergeTreeThread"; }
 
-    ~MergeTreeThreadSelectBlockInputProcessor() override;
+    ~MergeTreeThreadSelectProcessor() override;
 
 protected:
     /// Requests read task from MergeTreeReadPool and signals whether it got one
