+++ src/Storages/ya.make
@@ -16,8 +16,8 @@ SRCS(
     ColumnsDescription.cpp
     ConstraintsDescription.cpp
     Distributed/DirectoryMonitor.cpp
-    Distributed/DistributedBlockOutputStream.cpp
     Distributed/DistributedSettings.cpp
+    Distributed/DistributedSink.cpp
     IStorage.cpp
     IndicesDescription.cpp
     JoinSettings.cpp
@@ -41,7 +41,6 @@ SRCS(
     MergeTree/MergeAlgorithm.cpp
     MergeTree/MergeList.cpp
     MergeTree/MergeTreeBaseSelectProcessor.cpp
-    MergeTree/MergeTreeBlockOutputStream.cpp
     MergeTree/MergeTreeBlockReadUtils.cpp
     MergeTree/MergeTreeData.cpp
     MergeTree/MergeTreeDataMergerMutator.cpp
@@ -87,7 +86,8 @@ SRCS(
     MergeTree/MergeTreeSelectProcessor.cpp
     MergeTree/MergeTreeSequentialSource.cpp
     MergeTree/MergeTreeSettings.cpp
-    MergeTree/MergeTreeThreadSelectBlockInputProcessor.cpp
+    MergeTree/MergeTreeSink.cpp
+    MergeTree/MergeTreeThreadSelectProcessor.cpp
     MergeTree/MergeTreeWhereOptimizer.cpp
     MergeTree/MergeTreeWriteAheadLog.cpp
     MergeTree/MergeType.cpp
@@ -99,7 +99,6 @@ SRCS(
     MergeTree/ReplicatedFetchList.cpp
     MergeTree/ReplicatedMergeTreeAddress.cpp
     MergeTree/ReplicatedMergeTreeAltersSequence.cpp
-    MergeTree/ReplicatedMergeTreeBlockOutputStream.cpp
     MergeTree/ReplicatedMergeTreeCleanupThread.cpp
     MergeTree/ReplicatedMergeTreeLogEntry.cpp
     MergeTree/ReplicatedMergeTreeMergeStrategyPicker.cpp
@@ -108,6 +107,7 @@ SRCS(
     MergeTree/ReplicatedMergeTreePartHeader.cpp
     MergeTree/ReplicatedMergeTreeQueue.cpp
     MergeTree/ReplicatedMergeTreeRestartingThread.cpp
+    MergeTree/ReplicatedMergeTreeSink.cpp
     MergeTree/ReplicatedMergeTreeTableMetadata.cpp
     MergeTree/SimpleMergeSelector.cpp
     MergeTree/TTLMergeSelector.cpp
